Eliminate compiler warnings across LibTorchSharp and NativeAcceleration
- Remove #pragma once from 8 .cpp files
- Replace deprecated torch::IntList with at::IntArrayRef
- Fix NULL used for non-pointer types (cufftHandle, float returns)
- Remove unused variables and dead code
- Fix unsigned arithmetic bugs in tmax/tmin clamping (Xray.cu, Backproject.cu)
- Fix sign conversion in Remap.cu sentinel assignment
- Suppress #20054-D for benign __shared__ struct initialization
- Remove unnecessary device_functions.h include
- Replace deprecated cudaMemcpyToArray with cudaMemcpy2DToArray
- Remove Clang-only -Wno-unused-local-typedef from CMakeLists.txt

+++ NativeAcceleration/gtom/src/Relion/Backproject.cu
@@ -271,8 +271,8 @@ namespace gtom
 				}
 
 				position.x = tmin(dimvolume / 2, position.x);
-				position.y = tmax(0, tmin(dimvolume - 1, position.y));
-				position.z = tmax(0, tmin(dimvolume - 1, position.z));
+				position.y = tmax(0, tmin((int)dimvolume - 1, position.y));
+				position.z = tmax(0, tmin((int)dimvolume - 1, position.z));
 
 				float interpweight = __half2float(interpw[i]);
 				if (squareinterpweights)
@@ -306,8 +306,8 @@ namespace gtom
 					}
 
 					position.x = tmin(dimvolume / 2, position.x);
-					position.y = tmax(0, tmin(dimvolume - 1, position.y));
-					position.z = tmax(0, tmin(dimvolume - 1, position.z));
+					position.y = tmax(0, tmin((int)dimvolume - 1, position.y));
+					position.z = tmax(0, tmin((int)dimvolume - 1, position.z));
 
 					if (d_volumeft != NULL)
 					{
